Intial commit from vanilla 1.3.5 installation.
It appears as though `[cmder_root]/config/user-ConEmu.xml` is where all
customized settings are copied to once `cmder` exits. The canonical and
original source is `[cmder_root]/config/ConEmu.xml` but this gets copied
on first launch into `[cmder_root/vendor/conemu-maximus5/ConEmu.xml` and
the active configuration which is saved to whenever the settings dialog
Save action is invoked. A bit convoluted and confusing, but I think
that's the deal.

Source:

https://github.com/cmderdev/cmder#upgrading